refactor(game): move board header
